Split the 1200+ line mime_responds_test into 3 more focused and manageable test files.
